refactor: fix files names to be CamalCase.
