+++ lib/AppProvider.dart
@@ -1,6 +1,6 @@
 import 'package:chat_now/add_room/RoomComponent.dart';
 import 'package:flutter/material.dart';
-import 'model/user.dart';
+import 'model/User.dart';
 class AppProvider extends ChangeNotifier{
   User? currentuser; // nullable
   late List<RoomComponent> roomList=List.empty(growable: true); //save user's room list
