+++ lib/home/HomeScreen.dart
@@ -1,10 +1,10 @@
 import 'package:chat_now/add_room/RoomComponent.dart';
-import 'package:chat_now/add_room/addroom.dart';
+import 'package:chat_now/add_room/AddRoom.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../Background.dart';
-import '../appprovider.dart';
+import '../AppProvider.dart';
 
 class HomeScreen extends StatefulWidget {
   static const String routeName = 'home';
@@ -29,7 +29,10 @@ class _HomeScreenState extends State<HomeScreen> {
           length: 2,
           initialIndex: 0,
           child: Scaffold(
-            resizeToAvoidBottomInset: false,
+            drawer: Drawer(
+              elevation: 20,
+              child: Container(),
+            ),
             backgroundColor: Colors.transparent,
             appBar: AppBar(
               bottom: TabBar(indicatorColor: Colors.white, tabs: <Widget>[
@@ -40,7 +43,6 @@ class _HomeScreenState extends State<HomeScreen> {
                   child: Text('Browse'),
                 ),
               ]),
-              automaticallyImplyLeading: false,
               title: Text(
                 'Chat App',
                 style: TextStyle(fontSize: 25),
@@ -53,10 +55,10 @@ class _HomeScreenState extends State<HomeScreen> {
               children: <Widget>[
                 Center(
                   child:
-                  // provider.listSize() == 0
-                  //     ? Text("No rooms joined yet")
-                  //     :
-                  createRoomList(), //when the list is empty type no rooms joined
+                      // provider.listSize() == 0
+                      //     ? Text("No rooms joined yet")
+                      //     :
+                      createRoomList(), //when the list is empty type no rooms joined
                 ),
                 Center(
                   child: Text("Suggested rooms"),
@@ -86,18 +88,17 @@ class _HomeScreenState extends State<HomeScreen> {
               return Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: RoomComponent(new RoomData(
-                    roomImagePath: provider.roomList.elementAt(index).room.roomImagePath,
+                    roomImagePath:
+                        provider.roomList.elementAt(index).room.roomImagePath,
                     roomName: provider.roomList.elementAt(index).room.roomName,
                     members: provider.roomList.elementAt(index).room.members,
                     category: provider.roomList.elementAt(index).room.category,
-                    description: provider.roomList.elementAt(index).room.description
-                )
-                ),
+                    description:
+                        provider.roomList.elementAt(index).room.description)),
               );
             } else
               return Text("No rooms joined yet");
           }
-        }
-    );
+        });
   }
 }
